// src/modules/stationary-products/product.interface.ts
// 1. Create an interface representing a document in MongoDB.
export interface StationeryProduct {
  name: string;
  brand: string;
  price: number;
  category:
    | 'Writing'
    | 'Office Supplies'
    | 'Art Supplies'
    | 'Educational'
    | 'Technology';
  description: string;
  quantity: number;
  inStock: boolean;
}